+++ NEWS.md
@@ -1,6 +1,7 @@
 ## 2026-XX-XX v1.23.0-SNAPSHOT
 * Replace `mod-configuration` with `mod-settings` to get language and locale settings ([MODTEMPENG-117](https://folio-org.atlassian.net/browse/MODTEMPENG-117))
 * Upgrade module to Vert.x 5.0 and Java 21 ([MODTEMPENG-122](https://folio-org.atlassian.net/browse/MODTEMPENG-122))
+* Migrate to /locale API for tenant locale settings ([MODTEMPENG-133](https://folio-org.atlassian.net/browse/MODTEMPENG-133))
 
 ## 2025-03-13 v1.22.0
 * Support barcode image generate for HRID tokens (MODTEMPENG-111)